refactor: move code to ckpt_utils.py
